requirements.txt
Flask==1.1.1
mysql-connector-python==8.0.18
ocrspace==2.2
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.7
future
MySQL-python